fix : bookmarks user photo
