+++ src/Components/index.js
@@ -14,3 +14,4 @@ export { UserPosts } from "./UserPosts/UserPosts";
 export { EditProfileModal } from "./EditProfileModal/EditProfileModal";
 export { Loader } from "./Loader/Loader";
 export { EditPostModal } from "./EditPostModal/EditPostModal";
+export { BookmarkPost } from "./BookmarkPost/BookmarkPost";
